Separate Delta class
