+++ spec/lib/delta_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe Dydx:Delta do
+end
